Group player settings into its own controller
