+++ BookPlayer/Settings/SettingsViewController.swift
@@ -14,17 +14,12 @@ import UIKit
 
 class SettingsViewController: UITableViewController, MFMailComposeViewControllerDelegate {
     @IBOutlet weak var autoplayLibrarySwitch: UISwitch!
-    @IBOutlet weak var smartRewindSwitch: UISwitch!
-    @IBOutlet weak var boostVolumeSwitch: UISwitch!
-    @IBOutlet weak var globalSpeedSwitch: UISwitch!
     @IBOutlet weak var disableAutolockSwitch: UISwitch!
-    @IBOutlet weak var rewindIntervalLabel: UILabel!
-    @IBOutlet weak var forwardIntervalLabel: UILabel!
 
-    let supportSection: Int = 7
-    let githubLinkPath = IndexPath(row: 0, section: 7)
-    let supportEmailPath = IndexPath(row: 1, section: 7)
-    let siriShortcutPath = IndexPath(row: 0, section: 5)
+    let supportSection: Int = 4
+    let githubLinkPath = IndexPath(row: 0, section: 4)
+    let supportEmailPath = IndexPath(row: 1, section: 4)
+    let siriShortcutPath = IndexPath(row: 0, section: 3)
 
     var version: String = "0.0.0"
     var build: String = "0"
@@ -42,21 +37,11 @@ class SettingsViewController: UITableViewController, MFMailComposeViewController
         super.viewDidLoad()
 
         self.autoplayLibrarySwitch.addTarget(self, action: #selector(self.autoplayToggleDidChange), for: .valueChanged)
-        self.smartRewindSwitch.addTarget(self, action: #selector(self.rewindToggleDidChange), for: .valueChanged)
-        self.boostVolumeSwitch.addTarget(self, action: #selector(self.boostVolumeToggleDidChange), for: .valueChanged)
-        self.globalSpeedSwitch.addTarget(self, action: #selector(self.globalSpeedToggleDidChange), for: .valueChanged)
         self.disableAutolockSwitch.addTarget(self, action: #selector(self.disableAutolockDidChange), for: .valueChanged)
 
         // Set initial switch positions
         self.autoplayLibrarySwitch.setOn(UserDefaults.standard.bool(forKey: Constants.UserDefaults.autoplayEnabled.rawValue), animated: false)
-        self.smartRewindSwitch.setOn(UserDefaults.standard.bool(forKey: Constants.UserDefaults.smartRewindEnabled.rawValue), animated: false)
-        self.boostVolumeSwitch.setOn(UserDefaults.standard.bool(forKey: Constants.UserDefaults.boostVolumeEnabled.rawValue), animated: false)
-        self.globalSpeedSwitch.setOn(UserDefaults.standard.bool(forKey: Constants.UserDefaults.globalSpeedEnabled.rawValue), animated: false)
         self.disableAutolockSwitch.setOn(UserDefaults.standard.bool(forKey: Constants.UserDefaults.autolockDisabled.rawValue), animated: false)
-
-        // Retrieve initial skip values from PlayerManager
-        self.rewindIntervalLabel.text = self.formatDuration(PlayerManager.shared.rewindInterval)
-        self.forwardIntervalLabel.text = self.formatDuration(PlayerManager.shared.forwardInterval)
 
         guard
             let version = Bundle.main.infoDictionary!["CFBundleShortVersionString"] as? String,
@@ -69,47 +54,8 @@ class SettingsViewController: UITableViewController, MFMailComposeViewController
         self.build = build
     }
 
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        guard let viewController = segue.destination as? SkipDurationViewController else {
-            return
-        }
-
-        if segue.identifier == "AdjustRewindIntervalSegue" {
-            viewController.title = "Rewind"
-            viewController.selectedInterval = PlayerManager.shared.rewindInterval
-            viewController.didSelectInterval = { selectedInterval in
-                PlayerManager.shared.rewindInterval = selectedInterval
-
-                self.rewindIntervalLabel.text = self.formatDuration(PlayerManager.shared.rewindInterval)
-            }
-        }
-
-        if segue.identifier == "AdjustForwardIntervalSegue" {
-            viewController.title = "Forward"
-            viewController.selectedInterval = PlayerManager.shared.forwardInterval
-            viewController.didSelectInterval = { selectedInterval in
-                PlayerManager.shared.forwardInterval = selectedInterval
-
-                self.forwardIntervalLabel.text = self.formatDuration(PlayerManager.shared.forwardInterval)
-            }
-        }
-    }
-
     @objc func autoplayToggleDidChange() {
         UserDefaults.standard.set(self.autoplayLibrarySwitch.isOn, forKey: Constants.UserDefaults.autoplayEnabled.rawValue)
-    }
-
-    @objc func rewindToggleDidChange() {
-        UserDefaults.standard.set(self.smartRewindSwitch.isOn, forKey: Constants.UserDefaults.smartRewindEnabled.rawValue)
-    }
-
-    @objc func boostVolumeToggleDidChange() {
-        UserDefaults.standard.set(self.boostVolumeSwitch.isOn, forKey: Constants.UserDefaults.boostVolumeEnabled.rawValue)
-        PlayerManager.shared.boostVolume = self.boostVolumeSwitch.isOn
-    }
-
-    @objc func globalSpeedToggleDidChange() {
-        UserDefaults.standard.set(self.globalSpeedSwitch.isOn, forKey: Constants.UserDefaults.globalSpeedEnabled.rawValue)
     }
 
     @objc func disableAutolockDidChange() {
