+++ BookPlayer/Settings/ThemesViewController.swift
@@ -1,0 +1,39 @@
+//
+//  ThemesViewController.swift
+//  BookPlayer
+//
+//  Created by Gianni Carlo on 12/19/18.
+//  Copyright © 2018 Tortuga Power. All rights reserved.
+//
+
+import UIKit
+
+class ThemesViewController: UITableViewController {
+    let items = ["Light", "Dark"]
+
+    var selectedTheme: String!
+
+    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return self.items.count
+    }
+
+    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: "ThemeCell", for: indexPath)
+        let item = self.items[indexPath.row]
+
+        cell.textLabel?.text = item
+
+        if item == self.selectedTheme {
+            cell.accessoryType = .checkmark
+        }
+
+        return cell
+    }
+
+    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        let item = self.items[indexPath.row]
+
+        self.selectedTheme = item
+        self.tableView.reloadData()
+    }
+}
